chore(replay): colocate serialization with its steps, split log extraction
serializeSessionData moves out of the snappy recorder into the
serialize-session step, and console log extraction becomes its own
extract-console-logs step colocated with extractConsoleLogs. The
recorders keep only the data contracts (SerializedSessionData,
ExtractedConsoleLogs) and the aggregation methods.
